backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.2
cohere==4.27
qdrant-client==1.7.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pyjwt==2.8.0
openai==1.3.5